🔥 UPDATE: newsapi router functional
